autostream-agent-gemini/requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
streamlit
tiktoken
python-dotenv
pip-system-certs

